[Gold I] Title: 달빛 여우, Time: 916 ms, Memory: 67104 KB -BaekjoonHub

+++ 백준/Gold/16118. 달빛 여우/달빛 여우.java
@@ -13,9 +13,9 @@ class Node {
 
 class State {
     int node;
-    long totalDist;
+    int totalDist;
 
-    State(int node, long totalDist) {
+    State(int node, int totalDist) {
         this.node = node;
         this.totalDist = totalDist;
     }
@@ -23,10 +23,10 @@ class State {
 
 class WolfState {
     int node;
-    long totalDist;
+    int totalDist;
     int isFast;
 
-    WolfState(int node, long totalDist, int isFast) {
+    WolfState(int node, int totalDist, int isFast) {
         this.node = node;
         this.totalDist = totalDist;
         this.isFast = isFast;
@@ -58,9 +58,9 @@ public class Main {
         }
 
         // 1번에서 여우 출발하여 모든 노드로 가는 최단 경로 저장
-        long[] foxD = foxDijkstra(1);
+        int[] foxD = foxDijkstra(1);
         // 2번에서 늑대 출발하여 모든 노드로 가는 최단 경로 저장
-        long[] wolfD = wolfDijkstra(1);
+        int[] wolfD = wolfDijkstra(1);
 
         // 두 배열 비교 후 여우가 더 짧은 그루터기의 개수 출력
         int count = 0;
@@ -73,11 +73,11 @@ public class Main {
     }
 
     // O((E + V) log V) = 104000 * log4000 = 2백만 = 0.02초
-    static long[] foxDijkstra(int start) {
-        long[] d = new long[n + 1];
-        Arrays.fill(d, Long.MAX_VALUE);
-        PriorityQueue<State> pq = new PriorityQueue<>((s1, s2) -> Long.compare(s1.totalDist, s2.totalDist));
-        pq.offer(new State(start, 0L));
+    static int[] foxDijkstra(int start) {
+        int[] d = new int[n + 1];
+        Arrays.fill(d, Integer.MAX_VALUE);
+        PriorityQueue<State> pq = new PriorityQueue<>((s1, s2) -> Integer.compare(s1.totalDist, s2.totalDist));
+        pq.offer(new State(start, 0));
         d[start] = 0;
 
         while (!pq.isEmpty()) {
@@ -88,7 +88,7 @@ public class Main {
             }
 
             for (Node next : graph.get(cur.node)) {
-                long nextDist = d[cur.node] + (next.weight * 2);
+                int nextDist = d[cur.node] + (next.weight * 2);
                 if (d[next.to] > nextDist) {
                     d[next.to] = nextDist;
                     pq.offer(new State(next.to, nextDist));
@@ -99,14 +99,14 @@ public class Main {
     }
 
     // 처음 시행시 가중치 절반, 다음은 가중치 두배를 반복
-    static long[] wolfDijkstra(int start) {
+    static int[] wolfDijkstra(int start) {
         // 다음번에 빠르게 가는지, 아닌지에 대한 정보에 따라 상태 나누어서 저장, 0은 다음번에 빠르게 아님, 1은 다음번에 빠르게
-        long[][] d = new long[n + 1][2];
+        int[][] d = new int[n + 1][2];
         for (int i = 0; i <= n; i++) {
-            Arrays.fill(d[i], Long.MAX_VALUE);
+            Arrays.fill(d[i], Integer.MAX_VALUE);
         }
-        PriorityQueue<WolfState> pq = new PriorityQueue<>((s1, s2) -> Long.compare(s1.totalDist, s2.totalDist));
-        pq.offer(new WolfState(start, 0L, 1));
+        PriorityQueue<WolfState> pq = new PriorityQueue<>((s1, s2) -> Integer.compare(s1.totalDist, s2.totalDist));
+        pq.offer(new WolfState(start, 0, 1));
         d[start][1] = 0;
 
         while (!pq.isEmpty()) {
@@ -117,7 +117,7 @@ public class Main {
             }
 
             for (Node next : graph.get(cur.node)) {
-                long nextDist = cur.totalDist + (cur.isFast == 1 ? next.weight : next.weight * 4L);
+                int nextDist = cur.totalDist + (cur.isFast == 1 ? next.weight : next.weight * 4);
 
                 int toggledIsFast = cur.isFast == 1 ? 0 : 1;
                 if (d[next.to][toggledIsFast] > nextDist) {
@@ -127,7 +127,7 @@ public class Main {
             }
         }
 
-        long[] dist = new long[n + 1];
+        int[] dist = new int[n + 1];
         for (int i = 0; i <= n; i++) {
             dist[i] = Math.min(d[i][0], d[i][1]);
         }
